API documentation

Exported functions

SomervilleCouncilParser.display_items_by_meeting — Method

display_items_by_meeting(items::DataFrame)

Display items grouped by meeting, along with the meeting's date, name, and url.

SomervilleCouncilParser.filter_agenda — Method

filter_agenda(agenda::DataFrame, search_terms; case_invariant=true)

Return copy of agenda filtered such that only rows containing an item from search_terms in the description of the item will be preserved. When case_invariant is true, ignores uppercase vs lowercase letters.

SomervilleCouncilParser.get_agenda_items — Method

get_agenda_items(meeting_link; cache_dir=nothing)

Return DataFrame of agenda items from meeting meeting_link. meeting_link can be either the full link ("http://somervillecityma.iqm2.com/Citizens/Detail_Meeting.aspx?ID=2570") or the meeting ID (2570).

When cache_dir is not empty, agendas are read from the cache if they've been previously downloaded. If they do not exist in the cache, they'll be added to it.

SomervilleCouncilParser.request_meetings — Method

request_meetings(start_date, stop_date)

Return list of meetings that occurred (or will occur) between start_date and stop_date.

SomervilleCouncilParser.search_agendas_for_content — Method

search_agendas_for_content(start_date, stop_date, search_terms;
                           cache_dir=nothing, case_invariant=true)

Return NamedTuple containing meetings (a DataFrame of meetings found within the given start_date-stop_date range) and items (a DataFrame of all items containing at least one of the search_terms).

Non-exported functions

SomervilleCouncilParser.request_agenda_items — Method

request_agenda_items(meeting_id; verbose=false)

Return DataFrame of agenda items from meeting meeting_link. meeting_link can be either the full link ("http://somervillecityma.iqm2.com/Citizens/Detail_Meeting.aspx?ID=2570") or the meeting ID (2570).

SomervilleCouncilParser.validate_date — Method

validate_date(date::AbstractString)
validate_date(date::Date)

Returns month/day/year string in format accepted by city calendar. Valid input formats are "<M>/<D>/<YYYY>" (6/1/2017, 06/01/2017) or Date(yyyy,mm,dd) (Date(2017,6,1)).